Redesign with animated landing page
