+++ README.md
@@ -2,8 +2,9 @@
 
 مشروع لفك تشفير أكواد Node.js المشفرة باستخدام خوارزميات WebCrack محليًا.
 
-يتميز الإصدار الجديد بواجهة حديثة مع دعم الوضعين المظلم والفاتح، واستخدام محرر
-Monaco لكتابة الكود مع عرض التحليل المباشر لعدد الأسطر والمتغيرات والدوال.
+يتميز الإصدار الجديد بواجهة حديثة مع صفحة دخول متحركة.
+بعد الضغط على زر **ابدأ** ستنتقل إلى واجهة العمل التي تتضمن دعماً للوضعين المظلم والفاتح،
+ومحرر Monaco مع عرض التحليل المباشر لعدد الأسطر والمتغيرات والدوال.
 
 ## كيفية التشغيل محليًا
