+++ components/Logo.tsx
@@ -1,0 +1,18 @@
+export default function Logo() {
+  return (
+    <svg
+      width="80"
+      height="80"
+      viewBox="0 0 100 100"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <rect x="5" y="5" width="90" height="90" rx="10" fill="#0f172a" />
+      <path
+        d="M20 50h20l5-10 5 20 5-20 5 10h20"
+        stroke="#00ff9a"
+        strokeWidth="4"
+        fill="none"
+      />
+    </svg>
+  )
+}
